+++ Components/Essential/BasicDraw.cs
@@ -156,7 +156,7 @@ namespace OrbitVR.Components.Essential {
 
       Layers layer = parent.IsPlayer ? Layers.Player : DrawLayer;
         room.Camera.Draw(parent.body.texture, parent.body.pos, parent.body.color*(AlphaPercent/100f), parent.body.scale,
-                         parent.body.orient, layer);
+                         parent.body.orient, (int)layer);
     }
   }
 }
