+++ Framework/Input.cs
@@ -241,7 +241,7 @@ namespace OrbitVR.Framework {
       if (drawRing) {
         float scale = (radius*2f)/128;//Todo:Assets.TextureDict[Textures.Ring].Width;
         float alpha = (((float) Math.Sin(OrbIt.Game.Time.TotalGameTime.TotalMilliseconds/300f) + 1f)/4f) + 0.25f;
-        player.room.Camera.Draw(Textures.Ring, player.node.body.pos, player.pColor*alpha, scale, Layers.Under2);
+        player.room.Camera.Draw(Textures.Ring, player.node.body.pos, player.pColor*alpha, scale, (int)Layers.Under2);
       }
       return dir;
     }
