+++ Framework/Link.cs
@@ -478,7 +478,7 @@ namespace OrbitVR.Framework {
           VMath.NormalizeSafe(ref perp);
           perp *= 2;
 
-          room.Camera.DrawLine(source.body.pos, target.body.pos, 2f, color1, Layers.Under3);
+          room.Camera.DrawLine(source.body.pos, target.body.pos, 2f, color1, (int)Layers.Under3);
 
           //Utils.DrawLine(spritebatch, source.transform.position + perp, target.transform.position + perp, 2f, col, room);
           //Utils.DrawLine(spritebatch, source.transform.position - perp, target.transform.position - perp, 2f, col, room);
@@ -489,8 +489,8 @@ namespace OrbitVR.Framework {
           Vector2 center = (target.body.pos + source.body.pos)/2;
 
           Vector2 point = target.body.pos - (diff/5);
-          room.Camera.DrawLine(point + perp, target.body.pos, 2f, color1, Layers.Under3);
-          room.Camera.DrawLine(point - perp, target.body.pos, 2f, color1, Layers.Under3);
+          room.Camera.DrawLine(point + perp, target.body.pos, 2f, color1, (int)Layers.Under3);
+          room.Camera.DrawLine(point - perp, target.body.pos, 2f, color1, (int)Layers.Under3);
         }
       }
     }
