+++ Physics/Shape.cs
@@ -217,7 +217,7 @@ namespace OrbitVR.Physics {
 
         Vector2 v1 = position + a1;
         Vector2 v2 = position + a2;
-        body.room.Camera.DrawLine(v1, v2, LineThickness, color, Layers.Over2);
+        body.room.Camera.DrawLine(v1, v2, LineThickness, color, (int)Layers.Over2);
 
         if (RecurseDrawEnabled) {
           DrawRecurse(body.pos + a1, RecurseCount, 1f);
@@ -239,7 +239,7 @@ namespace OrbitVR.Physics {
 
         Vector2 v1 = pos + a1;
         Vector2 v2 = pos + a2;
-        body.room.Camera.DrawLine(v1, v2, LineThickness, body.color, Layers.Under5);
+        body.room.Camera.DrawLine(v1, v2, LineThickness, body.color, (int)Layers.Under5);
 
         //Draw(pos, count, scale, scalediff);
       }
@@ -258,7 +258,7 @@ namespace OrbitVR.Physics {
 
         Vector2 v1 = pos + a1;
         Vector2 v2 = pos + a2;
-        body.room.Camera.DrawLine(v1, v2, 1f, body.color, Layers.Under5);
+        body.room.Camera.DrawLine(v1, v2, 1f, body.color, (int)Layers.Under5);
 
         DrawRecurse(pos + a1, count, scale);
       }
